requirements.txt
Django==2.1.5
pytz==2018.9
psycopg2>=2.7,<3.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
boto3==1.9.83
django-storages==1.7.1
celery==3.1.18